Remove TODO for listing "channels subbed to"
The reason here is the need for a back-end:
https://twitchemotes.com/apidocs

Refer to the "Caching" section of that doc

Currently the only way (that I know of) to fetch a list of
subscriptions is to map emoticon set IDs to channels. There is a free
service that provides this, but it requires a proprietary server to
cache the mapping (the mapping is provided as a large flat file).

Sadly, until either 1) this app goes paid (cover server costs) or 2)
twitch open sources the mapping (a new API endpoint?), this feature is
blocked/back-logged indefinitely.